refactor(examples): refactor ngrx selectors and todos state
