+++ src/app/examples/examples.module.ts
@@ -3,32 +3,28 @@ import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { HttpClient } from '@angular/common/http';
 
 import { SharedModule } from '@app/shared';
+import { environment } from '@env/environment';
 
+import { FEATURE_NAME, reducers } from './examples.state';
 import { ExamplesRoutingModule } from './examples-routing.module';
 import { ExamplesComponent } from './examples/examples.component';
-import { TodosComponent } from './todos/todos.component';
-import { todosReducer } from './todos/todos.reducer';
+import { TodosComponent } from './todos/todos/todos.component';
 import { TodosEffects } from './todos/todos.effects';
 import { StockMarketComponent } from './stock-market/stock-market.component';
-import { stockMarketReducer } from './stock-market/stock-market.reducer';
 import { StockMarketEffects } from './stock-market/stock-market.effects';
 import { StockMarketService } from './stock-market/stock-market.service';
 import { ParentComponent } from './theming/parent/parent.component';
 import { ChildComponent } from './theming/child/child.component';
 import { AuthenticatedComponent } from './authenticated/authenticated.component';
-import { HttpClient } from '@angular/common/http';
-import { environment } from '@env/environment';
 
 @NgModule({
   imports: [
     SharedModule,
     ExamplesRoutingModule,
-    StoreModule.forFeature('examples', {
-      todos: todosReducer,
-      stocks: stockMarketReducer
-    }),
+    StoreModule.forFeature(FEATURE_NAME, reducers),
     TranslateModule.forChild({
       loader: {
         provide: TranslateLoader,
