+++ src/app/examples/todos/todos.actions.spec.ts
@@ -1,25 +1,32 @@
-import * as fromTodos from './todos.reducer';
+import {
+  ActionTodosAdd,
+  ActionTodosFilter,
+  ActionTodosPersist,
+  ActionTodosRemoveDone,
+  ActionTodosToggle,
+  TodosActionTypes
+} from './todos.actions';
 
 describe('Todos Actions', () => {
   describe('TodosAdd', () => {
     it('should create an action', () => {
-      const action = new fromTodos.ActionTodosAdd({ name: 'test' });
+      const action = new ActionTodosAdd({ name: 'test' });
       expect(action.payload).toEqual(
         jasmine.objectContaining({
           name: 'test'
         })
       );
-      expect(action.type).toEqual(fromTodos.TodosActionTypes.ADD);
+      expect(action.type).toEqual(TodosActionTypes.ADD);
       expect(action.payload.id).toBeDefined();
     });
   });
 
   describe('ActionTodosToggle', () => {
     it('should create an action', () => {
-      const action = new fromTodos.ActionTodosToggle({ id: '1' });
+      const action = new ActionTodosToggle({ id: '1' });
 
       expect({ ...action }).toEqual({
-        type: fromTodos.TodosActionTypes.TOGGLE,
+        type: TodosActionTypes.TOGGLE,
         payload: { id: '1' }
       });
     });
@@ -27,20 +34,20 @@ describe('Todos Actions', () => {
 
   describe('ActionTodosRemoveDone', () => {
     it('should create an action', () => {
-      const action = new fromTodos.ActionTodosRemoveDone();
+      const action = new ActionTodosRemoveDone();
 
       expect({ ...action }).toEqual({
-        type: fromTodos.TodosActionTypes.REMOVE_DONE
+        type: TodosActionTypes.REMOVE_DONE
       });
     });
   });
 
   describe('ActionTodosFilter', () => {
     it('should create an action', () => {
-      const action = new fromTodos.ActionTodosFilter({ filter: 'DONE' });
+      const action = new ActionTodosFilter({ filter: 'DONE' });
 
       expect({ ...action }).toEqual({
-        type: fromTodos.TodosActionTypes.FILTER,
+        type: TodosActionTypes.FILTER,
         payload: { filter: 'DONE' }
       });
     });
@@ -48,11 +55,13 @@ describe('Todos Actions', () => {
 
   describe('ActionTodosPersist', () => {
     it('should create an action', () => {
-      const action = new fromTodos.ActionTodosPersist({ todos: [] });
+      const action = new ActionTodosPersist({
+        todos: { filter: 'ALL', items: [] }
+      });
 
       expect({ ...action }).toEqual({
-        type: fromTodos.TodosActionTypes.PERSIST,
-        payload: { todos: [] }
+        type: TodosActionTypes.PERSIST,
+        payload: { todos: { filter: 'ALL', items: [] } }
       });
     });
   });
