+++ src/app/examples/todos/todos.effects.ts
@@ -5,11 +5,9 @@ import { tap } from 'rxjs/operators';
 
 import { LocalStorageService } from '@app/core';
 
-import {
-  ActionTodosPersist,
-  TODOS_KEY,
-  TodosActionTypes
-} from './todos.reducer';
+import { ActionTodosPersist, TodosActionTypes } from './todos.actions';
+
+export const TODOS_KEY = 'EXAMPLES.TODOS';
 
 @Injectable()
 export class TodosEffects {
