+++ src/app/examples/todos/todos.model.ts
@@ -1,0 +1,12 @@
+export interface Todo {
+  id: string;
+  name: string;
+  done: boolean;
+}
+
+export type TodosFilter = 'ALL' | 'DONE' | 'ACTIVE';
+
+export interface TodosState {
+  items: Todo[];
+  filter: TodosFilter;
+}
