+++ src/app/examples/todos/todos.reducer.ts
@@ -1,53 +1,7 @@
 import { v4 as uuid } from 'uuid';
-import { Action } from '@ngrx/store';
 
-export const TODOS_KEY = 'EXAMPLES.TODOS';
-
-export enum TodosActionTypes {
-  ADD = '[Todos] Add',
-  TOGGLE = '[Todos] Toggle',
-  REMOVE_DONE = '[Todos] Remove Done',
-  FILTER = '[Todos] Filter',
-  PERSIST = '[Todos] Persist'
-}
-
-export class ActionTodosAdd implements Action {
-  readonly type = TodosActionTypes.ADD;
-  readonly payload: { id: string; name: string };
-
-  constructor({ id = uuid(), name = '' }: { id?: string; name: string }) {
-    this.payload = { id, name };
-  }
-}
-
-export class ActionTodosToggle implements Action {
-  readonly type = TodosActionTypes.TOGGLE;
-
-  constructor(readonly payload: { id: string }) {}
-}
-
-export class ActionTodosRemoveDone implements Action {
-  readonly type = TodosActionTypes.REMOVE_DONE;
-}
-
-export class ActionTodosFilter implements Action {
-  readonly type = TodosActionTypes.FILTER;
-
-  constructor(readonly payload: { filter: TodosFilter }) {}
-}
-
-export class ActionTodosPersist implements Action {
-  readonly type = TodosActionTypes.PERSIST;
-
-  constructor(readonly payload: { todos: Todo[] }) {}
-}
-
-export type TodosActions =
-  | ActionTodosAdd
-  | ActionTodosToggle
-  | ActionTodosRemoveDone
-  | ActionTodosFilter
-  | ActionTodosPersist;
+import { TodosActions, TodosActionTypes } from './todos.actions';
+import { Todo, TodosState } from './todos.model';
 
 export const initialState: TodosState = {
   items: [
@@ -61,8 +15,6 @@ export const initialState: TodosState = {
   ],
   filter: 'ALL'
 };
-
-export const selectorTodos = state => state.examples.todos;
 
 export function todosReducer(
   state: TodosState = initialState,
@@ -103,17 +55,4 @@ export function todosReducer(
     default:
       return state;
   }
-}
-
-export interface Todo {
-  id: string;
-  name: string;
-  done: boolean;
-}
-
-export type TodosFilter = 'ALL' | 'DONE' | 'ACTIVE';
-
-export interface TodosState {
-  items: Todo[];
-  filter: TodosFilter;
 }
