+++ src/app/examples/todos/todos.selectors.ts
@@ -1,0 +1,8 @@
+import { createSelector } from '@ngrx/store';
+
+import { ExamplesState, selectExamples } from '../examples.state';
+
+export const selectTodos = createSelector(
+  selectExamples,
+  (state: ExamplesState) => state.todos
+);
