+++ src/styles.scss
@@ -21,7 +21,7 @@
 @import 'app/app.component.scss-theme';
 @import 'app/shared/big-input/big-input.component.scss-theme';
 @import 'app/static/about/about.component.scss-theme';
-@import 'app/examples/todos/todos.component.scss-theme';
+@import 'app/examples/todos/todos/todos.component.scss-theme';
 @import 'app/examples/stock-market/stock-market.component.scss-theme';
 @import 'app/examples/theming/parent/parent.component.scss-theme';
 @import 'app/examples/theming/child/child.component.scss-theme';
